Add bitmap.h header
